Fixed a bunch of issues:
* Fixed the duplicate progress bars at the end of transmission; the problem is
that float number comparison is error-prone; convert float fraction to int
percentage first so that std::endl is printed exactly once in the end.
* Converted progress struct to use OOP.
* Fixed the unresponsive sender side progress bar; the problem is that we don't
poll frequently enough; we used to poll only before we decide to send a repair
symbol; decouple the interval of polling from the interval of repair symbol
transmission.