Fixed two more issues:
* Duplicate progress bars because the number of symbols we get on the receiver
side may be larger than the number of source symbols in the file; so I change
the granularity of progress bar on the receiver side to block as well.
* Failure with 100M file; this is because the block size we derive using our
computeOptimalBlockSize method is not compatible with RFC 6330's algorithm;
in a nutshell, we cannot decide the number of blocks first and then derive the
number of symbols; it's the other way around; that algorithm is described in
RFC6330 page 8; I switch to try-based approach to get the smallest number of
symbols per block that make libRaptorQ happy.